requirements.txt
flask
flask-cors
flask-socketio
eventlet
mysql-connector-python
gunicorn
PyJWT
cloudinary